Separate classes into separate files
